% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost.R
\name{as.fmap.matrix}
\alias{as.fmap.matrix}
\title{Generates an XGBoost feature map based on feature data.}
\usage{
\method{as.fmap}{matrix}(matrix_X)
}
\arguments{
\item{matrix_X}{A "matrix" object with independent variables.}
}
\value{
A "data.frame" object.
}
\description{
Generates an XGBoost feature map based on feature data.
}
\examples{
data(iris)
iris.matrix = model.matrix(Species ~ . - 1, data = iris)
iris.fmap = as.fmap(iris.matrix)
}
