% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost.R
\name{as.fmap}
\alias{as.fmap}
\title{Dispatches execution to the most appropriate XGBoost feature map generation function.}
\usage{
as.fmap(x)
}
\arguments{
\item{x}{A dataset object.}
}
\description{
Dispatches execution to the most appropriate XGBoost feature map generation function.
}
