% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard.R
\name{as.scorecard}
\alias{as.scorecard}
\title{Converts a "glm" object to a "scorecard" object.}
\usage{
as.scorecard(glm, odds = 10, base_points = 500, pdo = 100)
}
\arguments{
\item{glm}{A "glm" object with binomial family link function.}

\item{odds}{Odds ratio at base odds.}

\item{base_points}{Points where odds ratio is defined.}

\item{pdo}{Points to double the odds.}
}
\value{
A "scorecard" object.
}
\description{
Converts a "glm" object to a "scorecard" object.
}

