% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoration.R
\name{decorate.default}
\alias{decorate.default}
\title{Decorates a model object with "preProcess" and "pmml_options" elements.}
\usage{
\method{decorate}{default}(x, preProcess = NULL, pmml_options = NULL, ...)
}
\arguments{
\item{x}{The model object.}

\item{preProcess}{A "train::preProcess" object.}

\item{pmml_options}{A list of model type-dependent PMML conversion options.}

\item{...}{Further arguments.}
}
\description{
Decorates a model object with "preProcess" and "pmml_options" elements.
}
