% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoration.R
\name{decorate.earth}
\alias{decorate.earth}
\title{Decorates an "earth" object with an "xlevels" element.}
\usage{
\method{decorate}{earth}(x, data, ...)
}
\arguments{
\item{x}{An "earth" object.}

\item{data}{The training dataset.}

\item{...}{Arguments to pass on to the "decorate.default" function.}
}
\description{
Decorates an "earth" object with an "xlevels" element.
}
