% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoration.R
\name{decorate.elm}
\alias{decorate.elm}
\title{Decorates an "elm" object with a "model" element.}
\usage{
\method{decorate}{elm}(x, data, ...)
}
\arguments{
\item{x}{An "elm" object.}

\item{data}{The training dataset.}

\item{...}{Arguments to pass on to the "decorate.default" function.}
}
\description{
Decorates an "elm" object with a "model" element.
}
