% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoration.R
\name{decorate.elmNN}
\alias{decorate.elmNN}
\title{Decorates an "elmNN" object with a "model" element.}
\usage{
\method{decorate}{elmNN}(x, data, ...)
}
\arguments{
\item{x}{An "elmNN" object.}

\item{data}{The training dataset.}

\item{...}{Arguments to pass on to the "decorate.default" function.}
}
\description{
Decorates an "elmNN" object with a "model" element.
}
