% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoration.R
\name{decorate.randomForest}
\alias{decorate.randomForest}
\title{Decorates a "randomForest" object with PMML conversion options.}
\usage{
decorate.randomForest(x, compact = FALSE, ...)
}
\arguments{
\item{x}{A "randomForest" object.}

\item{compact}{A flag controlling if decision trees should be transformed from binary splits (FALSE) to multi-way splits (TRUE) representation.}

\item{...}{Arguments to pass on to the "decorate.default" function.}
}
\description{
Decorates a "randomForest" object with PMML conversion options.
}
