% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoration.R
\name{decorate}
\alias{decorate}
\title{Dispatches execution to the most appropriate model decoration function.}
\usage{
decorate(x, ...)
}
\arguments{
\item{x}{A model object.}

\item{...}{Arguments to pass on to the selected function.}
}
\description{
Dispatches execution to the most appropriate model decoration function.
}
