% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoration.R
\name{decorate.svm.formula}
\alias{decorate.svm.formula}
\title{Decorates a "svm.formula" object with an "xlevels" element.}
\usage{
\method{decorate}{svm.formula}(x, data, ...)
}
\arguments{
\item{x}{A "svm.formula" object.}

\item{data}{The training dataset.}

\item{...}{Arguments to pass on to the "decorate.default" function.}
}
\description{
Decorates a "svm.formula" object with an "xlevels" element.
}
