% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoration.R
\name{decorate.train}
\alias{decorate.train}
\title{Decorates the final model of a "train" object with model type-dependent elements.}
\usage{
\method{decorate}{train}(x, ...)
}
\arguments{
\item{x}{A "train" object.}

\item{...}{Arguments to pass on to the "decorate.default" function.}
}
\description{
Decorates the final model of a "train" object with model type-dependent elements.
}
