% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoration.R
\name{decorate.WrappedModel}
\alias{decorate.WrappedModel}
\title{Decorates a "WrappedModel" object with "invert_levels" element.
Additionally, decorates the learned model with model type-dependent elements.}
\usage{
\method{decorate}{WrappedModel}(x, invert_levels = FALSE, ...)
}
\arguments{
\item{x}{A "WrappedModel" object.}

\item{invert_levels}{A flag indicating if the learned model should assume normal (FALSE) or inverted (TRUE) ordering of category values for the binary categorical target field.}

\item{...}{Arguments to pass on to the "decorate.default" function}
}
\description{
Decorates a "WrappedModel" object with "invert_levels" element.
Additionally, decorates the learned model with model type-dependent elements.
}
