% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost.R
\name{genFMap.data.frame}
\alias{genFMap.data.frame}
\title{Generates an XGBoost feature map based on feature data.}
\usage{
\method{genFMap}{data.frame}(df_X)
}
\arguments{
\item{df_X}{A "data.frame" object with independent variables.}
}
\value{
A "data.frame" object.
}
\description{
Generates an XGBoost feature map based on feature data.
}
\examples{
data(iris)
iris.df = iris[, 1:4]
iris.fmap = genFMap(iris.df)
}
