% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost.R
\name{genFMap}
\alias{genFMap}
\title{Dispatches execution to the most appropriate XGBoost feature map generation function.}
\usage{
genFMap(x)
}
\arguments{
\item{x}{A dataset object.}
}
\description{
Dispatches execution to the most appropriate XGBoost feature map generation function.
}
