% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verification.R
\name{verify.default}
\alias{verify.default}
\title{Enhances a model object with verification data.}
\usage{
verify.default(x, newdata)
}
\arguments{
\item{x}{A model object.}

\item{newdata}{The verification dataset.}
}
\description{
Enhances a model object with verification data.
}
