% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verification.R
\name{verify.glm}
\alias{verify.glm}
\title{Enhances a "glm" object with verification data.}
\usage{
verify.glm(x, newdata, precision = 1e-13, zeroThreshold = 1e-13)
}
\arguments{
\item{x}{A "glm" object.}

\item{newdata}{The verification dataset.}

\item{precision}{Maximal relative error.}

\item{zeroThreshold}{Maximal absolute error near the zero value.}
}
\description{
Enhances a "glm" object with verification data.
}
