% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verification.R
\name{verify}
\alias{verify}
\title{Dispatches execution to the most appropriate model verification function.}
\usage{
verify(x, newdata, ...)
}
\arguments{
\item{x}{A model object.}

\item{newdata}{The verification dataset.}

\item{...}{Arguments to pass on to the selected function.}
}
\description{
Dispatches execution to the most appropriate model verification function.
}
