% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verification.R
\name{verify.train}
\alias{verify.train}
\title{Enhances a "train" object with verification data.}
\usage{
\method{verify}{train}(x, newdata, precision = 1e-13,
  zeroThreshold = 1e-13, ...)
}
\arguments{
\item{x}{A "train" object.}

\item{newdata}{The verification dataset.}

\item{precision}{Maximal relative error.}

\item{zeroThreshold}{Maximal absolute error near the zero value.}

\item{...}{Arguments to pass on to the "predict.train" method.}
}
\description{
Enhances a "train" object with verification data.
}
