% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verification.R
\name{verify.xgb.Booster}
\alias{verify.xgb.Booster}
\title{Enhances an "xgb.Booster" object with verification data.}
\usage{
\method{verify}{xgb.Booster}(
  x,
  newdata,
  precision = 1e-06,
  zeroThreshold = 1e-06,
  response_name = NULL,
  response_levels = c(),
  ...
)
}
\arguments{
\item{x}{An "xgb.Booster" object.}

\item{newdata}{The verification dataset.}

\item{precision}{Maximal relative error.}

\item{zeroThreshold}{Maximal absolute error near the zero value.}

\item{response_name}{The name of the target field.}

\item{response_levels}{A list of category values for a categorical target field.}

\item{...}{Arguments to pass on to the "predict.xgb.Booster" method.}
}
\description{
Enhances an "xgb.Booster" object with verification data.
}
