% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost.R
\name{write.fmap}
\alias{write.fmap}
\title{Writes XGBoost feature map to a file.}
\usage{
write.fmap(fmap, file)
}
\arguments{
\item{fmap}{An XGBoost feature map as a "data.frame" object.}

\item{file}{A filesystem path to the result file.}
}
\description{
Writes XGBoost feature map to a file.
}
