% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoration.R
\name{decorate.glmnet}
\alias{decorate.glmnet}
\title{Decorates a "glmnet" object with a "lambda.s" element.}
\usage{
decorate.glmnet(x, lambda.s, ...)
}
\arguments{
\item{x}{A "glmnet" object.}

\item{lambda.s}{The best lambda value. Must be one of listed "glmnet$lambda" values.}

\item{...}{Arguments to pass on to the "decorate.default" function.}
}
\description{
Decorates a "glmnet" object with a "lambda.s" element.
}
