% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoration.R
\name{decorate.party}
\alias{decorate.party}
\title{Decorates a "party" object with a "predicted" element.}
\usage{
decorate.party(x, ...)
}
\arguments{
\item{x}{A "party" object.}

\item{...}{Arguments to pass on to the "decorate.default" function.}
}
\description{
Decorates a "party" object with a "predicted" element.
}
