% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost.R
\name{genFMap}
\alias{genFMap}
\title{Generates an XGBoost feature map based on feature data.}
\usage{
genFMap(df_X)
}
\arguments{
\item{df_X}{A "data.frame" object with independent variables.}
}
\value{
A "data.frame" object.
}
\description{
Generates an XGBoost feature map based on feature data.
}
\examples{
data(iris)
iris.fmap = genFMap(iris[, 1:4])
}
